refactor(page): update views to import Card from common/Card
