+++ apps/web/src/views/DashboardSalesPage.tsx
@@ -1,6 +1,6 @@
 import { Progress } from 'antd';
 
-import Card from '../components/common/Card';
+import { Card } from '@/components/common/card';
 
 export function DashboardSalesPage() {
   return (
