+++ apps/web/src/views/ECommercePage.tsx
@@ -1,4 +1,4 @@
-import Card from '../components/common/Card';
+import { Card } from '@/components/common/card';
 import type { OrderProcessingStatus } from '@salesops/shared';
 
 const defaultOrderStatus: OrderProcessingStatus = 'Created';
