+++ apps/web/src/views/InventoryAdjustmentPage.tsx
@@ -1,4 +1,4 @@
-import Card from '@/components/common/Card';
+import { Card } from '@/components/common/card';
 
 export function InventoryAdjustmentPage() {
   return (
